dvipdfmx.c: Cast as kpathsea_debug = (uint32_t)atoi(optarg). Since kpathsea_debug is unsigned int, the cast must be redundant. However a debian person says that he has a negative value for --kpathsea-debug -1, strangely.
git-svn-id: svn://tug.org/texlive/trunk/Build/source@38453 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/dvipdfm-x/dvipdfmx.c
@@ -461,7 +461,7 @@ do_args (int argc, char *argv[], const char *source)
       break;
 
     case 133: /* --kpathsea-debug */
-      if ((kpathsea_debug = atoi(optarg)) < 0)
+      if ((kpathsea_debug = (uint32_t)atoi(optarg)) < 0)
         ERROR("Invalid kpathsearch debugging flags specified: %s", optarg);
       break;
